用户注册登录,feign 熔断通信日志配置完成
