+++ shared-trad-common-serv/pom.xml
@@ -14,7 +14,6 @@
     <packaging>pom</packaging>
     <modules>
         <module>shared-trad-common-api</module>
-        <module>shared-trad-common-configuration</module>
         <module>shared-trad-common-tools</module>
     </modules>
