--- shared-trad-common-serv/shared-trad-common-configuration/src/main/java/cn/suparking/common/conf/config/properties/RedisLettucePoolProperties.java
@@ -1,18 +1,0 @@
-package cn.suparking.common.conf.config.properties;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "spring.redis.lettuce.pool")
-public class RedisLettucePoolProperties {
-    private Integer maxActive;
-
-    private Integer maxIdle;
-
-    private Integer minIdle;
-
-    private Integer maxWait;
-}
