--- shared-trad-common-serv/shared-trad-common-configuration/src/main/java/cn/suparking/common/conf/config/properties/RedisProperties.java
@@ -1,20 +1,0 @@
-package cn.suparking.common.conf.config.properties;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "spring.redis")
-public class RedisProperties {
-    private String host;
-
-    private int port;
-
-    private String password;
-
-    private int database;
-
-    private int timeout;
-}
