--- shared-trad-common-serv/shared-trad-common-configuration/src/main/java/cn/suparking/common/conf/config/properties/ThreadPoolProperties.java
@@ -1,19 +1,0 @@
-package cn.suparking.common.conf.config.properties;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "sparking.pool")
-public class ThreadPoolProperties {
-
-    private Integer coreSize;
-
-    private Integer maxSize;
-
-    private Integer keepAlive;
-
-    private Integer queueCapacity;
-}
