+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/configuration/properties/RedisLettucePoolProperties.java
@@ -1,0 +1,18 @@
+package cn.suparking.customer.configuration.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "spring.redis.lettuce.pool")
+public class RedisLettucePoolProperties {
+    private Integer maxActive;
+
+    private Integer maxIdle;
+
+    private Integer minIdle;
+
+    private Integer maxWait;
+}
