+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/configuration/properties/RedisProperties.java
@@ -1,0 +1,20 @@
+package cn.suparking.customer.configuration.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "spring.redis")
+public class RedisProperties {
+    private String host;
+
+    private int port;
+
+    private String password;
+
+    private int database;
+
+    private int timeout;
+}
