+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/configuration/properties/ThreadPoolProperties.java
@@ -1,0 +1,19 @@
+package cn.suparking.customer.configuration.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "sparking.pool")
+public class ThreadPoolProperties {
+
+    private Integer coreSize;
+
+    private Integer maxSize;
+
+    private Integer keepAlive;
+
+    private Integer queueCapacity;
+}
