+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/configuration/properties/WxProperties.java
@@ -1,0 +1,15 @@
+package cn.suparking.customer.configuration.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "sparking.wx")
+public class WxProperties {
+
+    private String appid;
+
+    private String secret;
+}
