+++ shared-trad-customer-serv/shared-trad-customer-impl/src/main/java/cn/suparking/customer/controller/user/service/impl/UserServiceImpl.java
@@ -3,21 +3,30 @@ package cn.suparking.customer.controller.user.service.impl;
 import cn.suparking.common.api.beans.SpkCommonResult;
 import cn.suparking.customer.controller.user.service.UserService;
 import cn.suparking.customer.feign.user.UserTemplateService;
-import cn.suparking.user.api.beans.MiniLoginDTO;
+import cn.suparking.customer.tools.ReactiveRedisUtils;
 import cn.suparking.user.api.beans.MiniRegisterDTO;
+import cn.suparking.user.api.beans.UserDTO;
+import cn.suparking.user.api.enums.RegisterType;
+import cn.suparking.user.api.enums.UserStatus;
 import cn.suparking.user.api.vo.PhoneInfoVO;
 import cn.suparking.user.api.vo.SessionVO;
 import cn.suparking.user.api.vo.UserVO;
+import com.alibaba.fastjson.JSON;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.catalina.User;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.Objects;
+import java.util.Optional;
 
 @Slf4j
 @Service
 public class UserServiceImpl implements UserService {
+
+    /**
+     * wx user info expired timeout.
+     */
+    private static final int WX_USER_EXPIRED_TIME = 86400;
 
     private final UserTemplateService userTemplateService;
 
@@ -33,19 +42,31 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public UserVO register(final MiniRegisterDTO miniRegisterDTO) {
-        UserVO userVO = null;
         //1.根据code获取openId和sessionKey
         SessionVO sessionVO = userTemplateService.getSessionKey(miniRegisterDTO.getCode());
-        if (Objects.nonNull(sessionVO)) {
+        return Optional.ofNullable(sessionVO).map(item -> {
             // 解析手机号码,然后注册用户
             PhoneInfoVO phoneInfoVO = userTemplateService.getPhoneInfo(miniRegisterDTO.getPhoneCode());
-            if (Objects.nonNull(phoneInfoVO)) {
-            }
-            return null;
-
-        }
-        // RegisterVO
-        return userVO;
+            return Optional.ofNullable(phoneInfoVO).map(phone -> {
+                // 调用用户创建接口准备落库
+                UserDTO userDTO = UserDTO.builder()
+                        .iphone(phone.getPhoneNumber())
+                        .miniOpenId(item.getOpenid())
+                        .unionId(item.getUnionId())
+                        .enabled(UserStatus.ACTIVE.getCode())
+                        .registerType(RegisterType.WECHAT_MINI.getCode())
+                        .build();
+                Integer count = userTemplateService.createSharedUser(userDTO);
+                if (Objects.nonNull(count) && count.equals(1)) {
+                    UserVO userVO = userTemplateService.getUserByOpenId(userDTO.getMiniOpenId());
+                    if (Objects.nonNull(userVO)) {
+                        opsValue(userVO);
+                    }
+                    return userVO;
+                }
+                return null;
+            }).orElse(null);
+        }).orElse(null);
     }
 
     /**
@@ -65,4 +86,19 @@ public class UserServiceImpl implements UserService {
         return userVO;
     }
 
+    /**
+     * save data.
+     * @param userVO {@link UserVO}
+     */
+    private void opsValue(final UserVO userVO) {
+        ReactiveRedisUtils.putValue(userVO.getMiniOpenId(), JSON.toJSONString(userVO), WX_USER_EXPIRED_TIME).subscribe(
+            flag -> {
+                if (flag) {
+                    log.info("Key= " + userVO.getMiniOpenId() + " save redis success!");
+                } else {
+                    log.info("Key= " + userVO.getMiniOpenId() + " save redis failed!");
+                }
+            }
+        );
+    }
 }
