+++ shared-trad-user-serv/shared-trad-user-impl/src/main/java/cn/suparking/user/configuration/properties/ThreadPoolProperties.java
@@ -1,0 +1,19 @@
+package cn.suparking.user.configuration.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "sparking.pool")
+public class ThreadPoolProperties {
+
+    private Integer coreSize;
+
+    private Integer maxSize;
+
+    private Integer keepAlive;
+
+    private Integer queueCapacity;
+}
