+++ shared-trad-user-serv/shared-trad-user-impl/src/main/java/cn/suparking/user/service/UserServiceImpl.java
@@ -1,9 +1,9 @@
 package cn.suparking.user.service;
 
-import cn.suparking.common.tools.middles.ReactiveRedisUtils;
 import cn.suparking.user.api.beans.UserDTO;
 import cn.suparking.user.dao.entity.UserDO;
 import cn.suparking.user.dao.mapper.UserMapper;
+import cn.suparking.user.tools.ReactiveRedisUtils;
 import cn.suparking.user.vo.UserVO;
 import cn.suparking.user.service.intf.UserService;
 import com.alibaba.fastjson.JSON;
